java files including the folders from eclipse
